Ironing
Ironed out potential issues and exploits.
Removed unneeded command groups
Added filtered messages
Updated descriptions

+++ titlecase/titlecase.py
@@ -12,8 +12,6 @@ class TitleCase(commands.Cog):
     def __init__(self, bot):
         self.bot = bot
 
-    @commands.group()
-    @checks.admin_or_permissions(administrator=True)
     async def titlecase(self, ctx):
         """TitleCase settings"""
         if ctx.invoked_subcommand is None:
@@ -33,10 +31,12 @@ class TitleCase(commands.Cog):
         user = m.author.mention
         trigger = str(c(m.content))
         quote = trigger[2:-2]
+        if await self.bot.is_automod_immune(m):
+            return
         if (await self.bot.get_context(m)).valid or m.content.startswith('"') or m.content.startswith('\\'):
             return
         if m.author.bot is False and trigger != "[]":
-            await m.channel.send(f"{user} wrote *\"{quote}...\"* \
+            await self.bot.send_filtered(m.channel, f"{user} wrote *\"{quote}...\"* \
                                     \nPlease refrain from using weird capitals.", delete_after=30)
             try:
                 await m.delete()
